refactor: separate tRPC procedures into individual files
