+++ packages/api/index.ts
@@ -1,0 +1,6 @@
+
+export { appRouter, type AppRouter } from './trpc/router';
+export { sendEmail } from './modules/email/procedures/send-email';
+export { authRouter } from './modules/auth/procedures';
+export { postsRouter } from './modules/posts/procedures';
+export { usersRouter } from './modules/users/procedures';
