--- packages/api/index.tsx
@@ -1,5 +1,0 @@
-
-export { appRouter, type AppRouter } from './trpc/router';
-export { sendEmail } from './modules/email/procedures/send-email';
-export { verifyCodeRouter } from './modules/auth/procedures/verify-code';
-export { authRouter } from './modules/auth/procedures/register';
