+++ packages/api/modules/auth/procedures/register.ts
@@ -12,32 +12,7 @@ const userDataSchema = z.object({
   birthdate: z.string().transform(str => new Date(str)),
 });
 
-type UserData = z.infer<typeof userDataSchema>;
-
-export const authRouter = router({
-  checkEmailExists: publicProcedure
-    .input(z.object({
-      email: z.string().email(),
-    }))
-    .query(async ({ input, ctx }) => {
-      try {
-        const normalizedEmail = input.email.trim().toLowerCase();
-        
-        const existingUser = await ctx.prisma.user.findUnique({
-          where: { email: normalizedEmail },
-          select: { id: true }
-        });
-        
-        return { 
-          exists: !!existingUser,
-          email: normalizedEmail
-        };
-      } catch (error) {
-        console.error('Check email exists error:', error);
-        return { exists: false, email: input.email };
-      }
-    }),
-
+export const registerRouter = router({
   register: publicProcedure
     .input(userDataSchema)
     .mutation(async ({ input, ctx }) => {
@@ -88,82 +63,6 @@ export const authRouter = router({
       } catch (error) {
         console.error('Registration error:', error);
         return { success: false, error: 'Failed to create user' };
-      }
-    }),
-
-  getUser: publicProcedure
-    .input(z.string())
-    .query(async ({ input, ctx }) => {
-      try {
-        const user = await ctx.prisma.user.findFirst({
-          where: {
-            OR: [
-              { email: input },
-              { username: input },
-              { id: input }
-            ]
-          },
-          select: {
-            id: true,
-            firstName: true,
-            lastName: true,
-            email: true,
-            username: true,
-            isVerified: true,
-          }
-        });
-        
-        return user;
-      } catch (error) {
-        console.error('Get user error:', error);
-        return null;
-      }
-    }),
-
-  login: publicProcedure
-    .input(z.object({
-      emailOrPhone: z.string(),
-      password: z.string(),
-    }))
-    .mutation(async ({ input, ctx }) => {
-      try {
-        const user = await ctx.prisma.user.findFirst({
-          where: {
-            OR: [
-              { email: input.emailOrPhone },
-              { username: input.emailOrPhone }
-            ]
-          }
-        });
-
-        if (!user) {
-          return { success: false, error: 'User not found' };
-        }
-
-        const isValidPassword = await bcrypt.compare(input.password, user.password);
-        
-        if (!isValidPassword) {
-          return { success: false, error: 'Invalid password' };
-        }
-
-        if (!user.isVerified) {
-          return { success: false, error: 'Please verify your email before signing in' };
-        }
-
-        return {
-          success: true,
-          user: {
-            id: user.id,
-            firstName: user.firstName,
-            lastName: user.lastName,
-            email: user.email,
-            username: user.username,
-            isVerified: user.isVerified,
-          }
-        };
-      } catch (error) {
-        console.error('Login error:', error);
-        return { success: false, error: 'Login failed' };
       }
     }),
 });
