+++ packages/api/package.json
@@ -1,8 +1,8 @@
 {
     "name": "@repo/api",
     "version": "0.0.0",
-    "main": "./index.tsx",
-    "types": "./**/.tsx",
+    "main": "./index.ts",
+    "types": "./**/.ts",
     "scripts": {
         "type-check": "tsc --noEmit"
     },
